feat(authority): add durable, fail-closed revocation records
Implements RFC 0003 Amendment 1 part (c): revoke_token/revoke_approval
publish exclusive-create records under revoked-tokens/revoked-approvals
and report a three-way outcome (Revoked / AlreadyRevoked /
RevokedAfterConsumption). consume_bundle's existing revocation checks
(steps 2 and 6) and the legacy consume_token/consume_approval (defense
in depth) now observe real revocations via a shared check_not_revoked
helper that fails closed on a corrupt record. purge_expired purges the
two new directories on the same expiry rule as everything else.

Split the crate's inline test module into src/tests.rs first (pure
move, privacy-preserving) to keep lib.rs under the file-size gate
before adding the new code.
