+++ docs/backlog.md
@@ -204,7 +204,7 @@ repo audit; every entry below points at verified, real state of the code.
   Amendment 1 part (e) in `rfcs/0003-signed-approval-evidence.md` — use them
   verbatim.
 
-- [ ] **P1 | `crates/authority/` | Add durable, fail-closed revocation records. IN PROGRESS (2026-08-27)**
+- [x] **P1 | `crates/authority/` | Add durable, fail-closed revocation records.**
   Blocked until the amendment lands; claim only after the transaction entry
   above is checked off. Per amendment part (c): `revoke_*` APIs write
   exclusive-create records under the store root; consumption checks
@@ -999,3 +999,4 @@ repo audit; every entry below points at verified, real state of the code.
 - 2026-08-26: landed source-closure gate v1 for `crates/vault-v2-engine` (the tagged AST-gate item, re-sliced): new `tests/gate.rs` machinery + `tests/ast_gate.rs` config/teeth, dev-dep pins syn `=2.0.118` / proc-macro2 `=1.0.106` (both already in the lock). The real-crate test pins the exact six-file closure and passes; 20 teeth tests prove every rejection fires against fixture crates in tempdirs (orphan, `#[path]`, unadmitted include, unsafe hidden in macro tokens, `macro_rules!`, unsafe/extern outside the boundary, ambiguous/missing/cfg-disabled modules, denied attribute/derive/macro, smuggled `#[path …]` in token trees, raw identifiers, symlinks) plus the positive `FFI_BOUNDARY_FILES` admission path the FFI item will use. Teeth additionally verified by a real-tree mutation: an orphan `src/stray.rs` — invisible to the compiler with all auto-discovery off — failed the gate, and removal restored green. The exactly-two-entry-points proof, five `tests/ui/` fixtures, and macro-definition mutation tests need the engine API to exist, so they became a follow-up `needs:fable` entry ordered after the FFI/profile items, whose entries now carry the FFI_BOUNDARY_FILES/ROOTS coupling instructions. Gate ALL GREEN.
 - 2026-08-23: added table-driven tests in the new `crates/policy/tests/v2_rejection.rs` covering every `PolicyV2Error` variant: `InvalidContext` for nil `session_id`/`idempotency_key` and malformed `audience`/`venture_id`/`subject_id` (empty, untrimmed, >512 chars, control char), plus `MissingPrimaryResource` from `evaluate_prepared` against a `PreparedInvocation` built from a manifest with an empty `resource_bindings` array. No production code changed — this closed a coverage gap only. Reused the signed-manifest fixture pattern from `crates/capability/tests/capability_v2.rs`; added `sovereign-identity` and `serde_json_canonicalizer` as dev-dependencies of `crates/policy` to build it. Teeth verified by two temporary mutations, both caught and both reverted: dropping the `session_id.is_nil()` check, and replacing `MissingPrimaryResource` with a stub default. `cargo test -p sovereign-policy` now runs 5 tests, all passing. Full gate ALL GREEN (gate-self-test, file-size, fmt, clippy(workspace), test(workspace), tsc(frontend)).
 - 2026-08-26: implemented RFC 0003 Amendment 1's bundle transaction in `crates/authority` — `AuthorityStore::consume_bundle` claims a token, an approval, and an idempotency key as one recoverable transaction instead of three separate claims, per Amendment 1 part (a)/(b) exactly. Bundle id is a domain-separated SHA-256 over the three subject ids plus the invocation fingerprint (deterministic, so a same-input retry resumes its own bundle instead of colliding or burning). `AuthorityRecord` gained `#[serde(default)] bundle_hex: Option<String>` (old records keep parsing). The six protocol steps (intent, revocation pre-check, token claim, idempotency bind, approval claim, revocation re-check + commit) are private methods exposed to the in-crate test module so crash points can be driven directly, exactly as the amendment requires; `claim()` was refactored onto a new generic `publish_record` helper shared with the bundle-record publish path (no behavior change, removed duplication). `purge_expired` now also purges `bundles/` entries by the same expiry rule via a small `Expiring` trait, satisfying the amendment's "stale partial claims self-expire with their subjects." The revocation pre-check/re-check against `revoked-tokens/`/`revoked-approvals/` is implemented as specified in part (a), but nothing publishes into those directories yet — `revoke_token`/`revoke_approval` are the next, still-blocked backlog entry (part c); today the check is a no-op that will start observing real revocations once that entry lands. The existing single-claim methods (`consume_token`/`consume_approval`/`bind_idempotency`) are unchanged and stay for the `crates/capability` migration's own round, per the entry's scope note. All six pinned Amendment-1-part-(e) test names pass verbatim: `a_retried_bundle_after_any_interruption_completes_without_burning_claims`, `racing_bundles_over_the_same_token_have_exactly_one_winner`, `racing_retries_of_the_same_bundle_authorize_exactly_once`, `a_reopened_store_answers_a_partial_bundle_identically`, `a_foreign_uncommitted_bundle_denies_other_consumers`, `purge_removes_expired_bundles_and_their_claims`. `cargo test -p sovereign-authority` runs 14 tests, all passing (8 pre-existing plus the 6 above). Added `sha2 = { workspace = true }` as a dependency (already pinned workspace-wide; only `Cargo.lock`'s dependency graph changed). Full gate ALL GREEN (gate-self-test, file-size, fmt, clippy(workspace), test(workspace), tsc(frontend)).
+- 2026-08-27: implemented RFC 0003 Amendment 1 part (c), durable revocation, in `crates/authority`. `AuthorityStore` gained `revoked_tokens`/`revoked_approvals` directories (already-named constants, now actually opened and populated) and `revoke_token`/`revoke_approval`, both routing through one private `revoke()` that publishes an exclusive-create record via the existing `publish_record` primitive and returns a three-way `RevocationOutcome` (`Revoked` / `AlreadyRevoked` / `RevokedAfterConsumption`), matching part (c) verbatim. "Already authorized" is judged precisely, not just "claim exists": a new `subject_is_authorized` helper treats a legacy claim (no `bundle_hex`) as immediately authorized, but a bundle-claimed subject only once its bundle's `.committed` marker exists — a token claimed by an in-flight, uncommitted bundle is not yet an authorized effect, so revoking it reports the clean `Revoked` outcome, not `RevokedAfterConsumption`; this distinction is exercised directly in `revoking_a_consumed_claim_reports_the_distinct_outcome`. `bundle_check_revocation` (steps 2 and 6) and the legacy `consume_token`/`consume_approval` (defense in depth) now share one `check_not_revoked` free function that fails closed on both a present record and an unparsable one (`CorruptRecord`, never silently allowed). `purge_expired` purges the two revocation directories on the same expiry rule as everything else. All four pinned Amendment-1-part-(e) revocation test names pass verbatim: `a_revoked_token_fails_closed_across_reopen`, `revoking_a_consumed_claim_reports_the_distinct_outcome`, `a_revoke_vs_consume_race_ends_in_one_durable_outcome`, `a_corrupt_revocation_record_fails_closed`; the race test asserts the outcome pairing is exhaustively one of exactly two combinations (bundle authorized + `RevokedAfterConsumption`, or bundle denied `Revoked` + revoke reports `Revoked`) across 20 randomized runs, never any other pairing. `cargo test -p sovereign-authority` runs 18 tests, all passing. Before adding this code, moved the crate's `#[cfg(test)] mod tests { … }` out of `lib.rs` into `src/tests.rs` (declared as `#[cfg(test)] mod tests;`) as a pure, test-verified no-op refactor — `lib.rs` was at 1033 lines against the 1200 god-file limit and this round's addition would have pushed it over; see the new CLAUDE.md lesson. No new dependencies. Full gate ALL GREEN (gate-self-test, file-size, fmt, clippy(workspace), test(workspace), tsc(frontend)).
